Restructing brogue interface. Listeners to do.
